requirements.txt
-r requirements/build.txt
-r requirements/optional.txt
-r requirements/runtime.txt
-r requirements/tests.txt
m2r~=0.2.1
mmdeploy~=0.1.0
ncnn~=1.0.20211208
pyppl~=0.0.0
recommonmark~=0.7.1
sphinx~=4.0.2
tensorrt~=8.0.1.6

torchvision~=0.9.0
typing~=3.7.4.3
